+++ src/Features/LanguageServer/Protocol/Features/TodoComments/TodoCommentsListener.cs
@@ -34,7 +34,7 @@ namespace Microsoft.CodeAnalysis.TodoComments
         /// Remote service connection. Created on demand when we startup and then
         /// kept around for the lifetime of this service.
         /// </summary>
-        private RemoteServiceConnection<IRemoteTodoCommentsDiscoveryService>? _lazyConnection;
+        private RemoteServiceConnection<IRemoteTaskListService>? _lazyConnection;
 
         /// <summary>
         /// Queue where we enqueue the information we get from OOP to process in batch in the future.
@@ -94,12 +94,12 @@ namespace Microsoft.CodeAnalysis.TodoComments
 
             // Pass ourselves in as the callback target for the OOP service.  As it discovers
             // todo comments it will call back into us to notify VS about it.
-            _lazyConnection = client.CreateConnection<IRemoteTodoCommentsDiscoveryService>(callbackTarget: this);
+            _lazyConnection = client.CreateConnection<IRemoteTaskListService>(callbackTarget: this);
 
             // Now kick off scanning in the OOP process.
             // If the call fails an error has already been reported and there is nothing more to do.
             _ = await _lazyConnection.TryInvokeAsync(
-                (service, callbackId, cancellationToken) => service.ComputeTodoCommentsAsync(callbackId, cancellationToken),
+                (service, callbackId, cancellationToken) => service.ComputeTaskListItemsAsync(callbackId, cancellationToken),
                 cancellationToken).ConfigureAwait(false);
         }
