+++ src/Workspaces/Core/Portable/TaskList/IRemoteTaskListService.cs
@@ -9,39 +9,38 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Remote;
-using Microsoft.CodeAnalysis.TaskList;
 
-namespace Microsoft.CodeAnalysis.TodoComments
+namespace Microsoft.CodeAnalysis.TaskList
 {
     /// <summary>
     /// Interface to allow host (VS) to inform the OOP service to start incrementally analyzing and
     /// reporting results back to the host.
     /// </summary>
-    internal interface IRemoteTodoCommentsDiscoveryService
+    internal interface IRemoteTaskListService
     {
         internal interface ICallback
         {
-            ValueTask ReportTodoCommentDataAsync(RemoteServiceCallbackId callbackId, DocumentId documentId, ImmutableArray<TaskListItem> data, CancellationToken cancellationToken);
+            ValueTask ReportTaskListItemsAsync(RemoteServiceCallbackId callbackId, DocumentId documentId, ImmutableArray<TaskListItem> data, CancellationToken cancellationToken);
             ValueTask<TaskListOptions> GetOptionsAsync(RemoteServiceCallbackId callbackId, CancellationToken cancellationToken);
         }
 
-        ValueTask ComputeTodoCommentsAsync(RemoteServiceCallbackId callbackId, CancellationToken cancellation);
+        ValueTask ComputeTaskListItemsAsync(RemoteServiceCallbackId callbackId, CancellationToken cancellation);
         ValueTask ReanalyzeAsync(CancellationToken cancellationToken);
     }
 
-    [ExportRemoteServiceCallbackDispatcher(typeof(IRemoteTodoCommentsDiscoveryService)), Shared]
-    internal sealed class RemoteTodoCommentsDiscoveryCallbackDispatcher : RemoteServiceCallbackDispatcher, IRemoteTodoCommentsDiscoveryService.ICallback
+    [ExportRemoteServiceCallbackDispatcher(typeof(IRemoteTaskListService)), Shared]
+    internal sealed class RemoteTaskListCallbackDispatcher : RemoteServiceCallbackDispatcher, IRemoteTaskListService.ICallback
     {
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public RemoteTodoCommentsDiscoveryCallbackDispatcher()
+        public RemoteTaskListCallbackDispatcher()
         {
         }
 
         private ITaskListListener GetListener(RemoteServiceCallbackId callbackId)
             => (ITaskListListener)GetCallback(callbackId);
 
-        public ValueTask ReportTodoCommentDataAsync(RemoteServiceCallbackId callbackId, DocumentId documentId, ImmutableArray<TaskListItem> data, CancellationToken cancellationToken)
+        public ValueTask ReportTaskListItemsAsync(RemoteServiceCallbackId callbackId, DocumentId documentId, ImmutableArray<TaskListItem> data, CancellationToken cancellationToken)
             => GetListener(callbackId).ReportTaskListItemsAsync(documentId, data, cancellationToken);
 
         public ValueTask<TaskListOptions> GetOptionsAsync(RemoteServiceCallbackId callbackId, CancellationToken cancellationToken)
